Refactor chaining request mapper and request mapper support class into fili core
