+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/web/ChainingRequestMapperSpec.groovy
@@ -1,11 +1,9 @@
 // Copyright 2017 Yahoo Inc.
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
-package com.yahoo.bard.webservice.web.security
+package com.yahoo.bard.webservice.web
 
 import com.yahoo.bard.webservice.data.config.ResourceDictionaries
 import com.yahoo.bard.webservice.web.apirequest.ApiRequest
-import com.yahoo.bard.webservice.web.RequestMapper
-import com.yahoo.bard.webservice.web.RequestValidationException
 import com.yahoo.bard.webservice.web.apirequest.ApiRequestImpl
 
 import spock.lang.Specification
